Force 'path' type options to be strings
When running with `--no-cache`, it sets the 'cache' option to a boolean
'false', which was then coerced to a string by nopt.

We should allow users to specify any string, but `--no-...` makes it
_always_ be a boolean `false` which is surprising.  (Similarly if a path
is set to a `null` or some other non-string value in a .npmrc config
file.)

This causes non-string `'path'` type options to fail validation, so that
they'll fall back to the default when set to an invalid value.

PR-URL: https://github.com/npm/config/pull/11
Credit: @isaacs
Close: #11
Reviewed-by: @nlf