Remove email from auth set, support scoped _auth
The legacy CouchDB style registries require an email in order to
authenticate.  However, the email itself is never used as part of the
auth process after the initial account creation.  There is no need for
it to be scoped to a given registry or required in order to use _auth
with a base64 encoded username/password basic auth pair.

Also, add support for a registry-scoped _auth field.  I'm not sure why
that was removed in the first place, since npm-registry-fetch will parse
it, so we need to be moving a top-level _auth field under the default
registry, for all the same reasons we do so for tokens.